+++ lib/commercetools-api/src/Models/BusinessUnit/CompanyDraftModel.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 namespace Commercetools\Api\Models\BusinessUnit;
 
 use Commercetools\Api\Models\Common\BaseAddressCollection;
+use Commercetools\Api\Models\Customer\CustomerGroupAssignmentDraftCollection;
 use Commercetools\Api\Models\Store\StoreResourceIdentifierCollection;
 use Commercetools\Api\Models\Type\CustomFieldsDraft;
 use Commercetools\Api\Models\Type\CustomFieldsDraftModel;
@@ -120,6 +121,12 @@ final class CompanyDraftModel extends JsonObjectModel implements CompanyDraft
      */
     protected $custom;
 
+    /**
+     *
+     * @var ?CustomerGroupAssignmentDraftCollection
+     */
+    protected $customerGroupAssignments;
+
 
     /**
      * @psalm-suppress MissingParamType
@@ -140,6 +147,7 @@ final class CompanyDraftModel extends JsonObjectModel implements CompanyDraft
         ?array $billingAddresses = null,
         ?int $defaultBillingAddress = null,
         ?CustomFieldsDraft $custom = null,
+        ?CustomerGroupAssignmentDraftCollection $customerGroupAssignments = null,
         ?string $unitType = null
     ) {
         $this->key = $key;
@@ -157,6 +165,7 @@ final class CompanyDraftModel extends JsonObjectModel implements CompanyDraft
         $this->billingAddresses = $billingAddresses;
         $this->defaultBillingAddress = $defaultBillingAddress;
         $this->custom = $custom;
+        $this->customerGroupAssignments = $customerGroupAssignments;
         $this->unitType = $unitType ?? self::DISCRIMINATOR_VALUE;
     }
 
@@ -490,6 +499,27 @@ final class CompanyDraftModel extends JsonObjectModel implements CompanyDraft
         return $this->custom;
     }
 
+    /**
+     * <p>Customer Groups to assign the Business Unit to.</p>
+     * <p>They are considered during <a href="/../api/pricing-and-discounts-overview#line-item-price-selection">line Item price selection</a>, if provided (non-null).</p>
+     *
+     *
+     * @return null|CustomerGroupAssignmentDraftCollection
+     */
+    public function getCustomerGroupAssignments()
+    {
+        if (is_null($this->customerGroupAssignments)) {
+            /** @psalm-var ?list<stdClass> $data */
+            $data = $this->raw(self::FIELD_CUSTOMER_GROUP_ASSIGNMENTS);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->customerGroupAssignments = CustomerGroupAssignmentDraftCollection::fromArray($data);
+        }
+
+        return $this->customerGroupAssignments;
+    }
+
 
     /**
      * @param ?string $key
@@ -609,5 +639,13 @@ final class CompanyDraftModel extends JsonObjectModel implements CompanyDraft
     public function setCustom(?CustomFieldsDraft $custom): void
     {
         $this->custom = $custom;
+    }
+
+    /**
+     * @param ?CustomerGroupAssignmentDraftCollection $customerGroupAssignments
+     */
+    public function setCustomerGroupAssignments(?CustomerGroupAssignmentDraftCollection $customerGroupAssignments): void
+    {
+        $this->customerGroupAssignments = $customerGroupAssignments;
     }
 }